Cleanup: duplication / layer violation: WebDAV props: relocate properties reading ({S|Des}erializeXml()) into new methods in adaptor class.
Reason being that the properties format is (supposed to be)
internal knowledge of that implementation
and thus is NOT supposed to be handled outside.

And since the adaptor method makes use of the internal provider.WriteFile() API,
make it public
(I believe that the reportUpdatedFile parameter *is* relevant
for the public interface of the SVN-interface-providing TFSSourceControlProvider).