PutHandler: fix improperly handled pseudo PUT within a COPY / PUT / DELETE sequence; adapt tests.
Subversion likes to do a dummy PUT after a COPY with actually unchanged
data. For proper handling, we need to calculate hashes to figure out
whether the data actually changed.

Also, add a large amount of corrections to unit tests
to properly drive the hugely corrected PUT handling.
While these tests do a PUT on an activity ID
*without* having executed MKACTIVITY before,
it seems that for test reasons
this is to be considered "ok".
Or perhaps better we should be adding base helpers
which take care of having a proper MKACTIVITY done
prior to executing any within-activity actions.